moving the performance runner into its own benchmark category
