+++ applications/stream/stream.cpp
@@ -1,6 +1,6 @@
 // stream.cpp: stream benchmarks of vector operations
 //
-// Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal number project, which is released under an MIT Open Source license.
 #include <universal/utility/directives.hpp>
@@ -31,9 +31,8 @@
 #define POSIT_THROW_ARITHMETIC_EXCEPTION 1
 #include <universal/number/posit/posit.hpp>
 
-#include <universal/verification/performance_runner.hpp>
-#include <universal/verification/test_status.hpp>
-#include <universal/verification/test_reporters.hpp>
+#include <universal/verification/test_suite.hpp>
+#include <universal/benchmark/performance_runner.hpp>
 
 template<typename Scalar>
 void Copy(std::vector<Scalar>& c, const std::vector<Scalar>& a, size_t start, size_t end) {
