+++ benchmark/performance/arithmetic/compare/performance.cpp
@@ -22,7 +22,7 @@
 // is representable
 #include <universal/functions/isrepresentable.hpp>
 #include <universal/verification/test_suite.hpp>
-#include <universal/verification/performance_runner.hpp>
+#include <universal/benchmark/performance_runner.hpp>
 
 template<typename NativeFloat>
 void CopyWorkload(size_t NR_OPS) {
