+++ benchmark/performance/arithmetic/fixpnt/performance.cpp
@@ -13,7 +13,7 @@
 // is representable
 #include <universal/functions/isrepresentable.hpp>
 #include <universal/verification/test_status.hpp> // ReportTestResult
-#include <universal/verification/performance_runner.hpp>
+#include <universal/benchmark/performance_runner.hpp>
 
 /*
    The goal of the arbitrary fixpnts is to provide a constrained big fixpnt type
