+++ internal/blocktriple/performance/performance.cpp
@@ -1,6 +1,6 @@
 //  performance.cpp : performance benchmarking for internal blocktriple operators
 //
-// Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <universal/utility/directives.hpp>
@@ -11,10 +11,9 @@
 #include <chrono>
 
 #include <universal/internal/blocktriple/blocktriple.hpp>
-#include <universal/verification/test_status.hpp>
-#include <universal/verification/test_reporters.hpp>
+#include <universal/verification/test_suite.hpp>
 #include <universal/verification/blockbinary_test_status.hpp>
-#include <universal/verification/performance_runner.hpp>
+#include <universal/benchmark/performance_runner.hpp>
 
 namespace sw::universal::internal {
