+++ static/cfloat/performance/perf.cpp
@@ -1,6 +1,6 @@
 //  perf.cpp : baseline performance benchmarking for cfloat arithmetic operators
 //
-// Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <universal/utility/directives.hpp>
@@ -11,9 +11,8 @@
 #include <chrono>
 
 #include <universal/number/cfloat/cfloat.hpp>
-#include <universal/verification/test_status.hpp>
-#include <universal/verification/test_reporters.hpp>
-#include <universal/verification/performance_runner.hpp>
+#include <universal/verification/test_suite.hpp>
+#include <universal/benchmark/performance_runner.hpp>
 
 namespace sw::universal::internal {
